Don't focus until scroll has completed
Focusing before the scrolling has been completed causes a momentary
flicker on screen (tested in Chrome 43). This is because calling
focus() instantly scrolls the page to the focused section, only for the
jQuery animate() call to kick into action a few ms later, jumping back
to the original starting location and *then* animating the scroll. Not
setting the focus until the scroll is completed resolves this issue.

This was introduced by pr #26.